Snowball Launcher
Snowball launcher works out of "OilPoint" GameObject using script attached to "OilPoint".
------------------------
Bugs - The oil from the oil powerup is not visible, however this bug was fixed before. This was fixed but never saved on my Mac so the changes were not pushed. The bug was never fixed. This is a minor bug however.